imglib2-scripting: integral histogram now accepts an arbitrary type rather than defaulting to, in most cases, the UnsignedAnyBitType which is very slow. The HistogramFeatures now takes an Img containing the integral histogram as argument, and also includes the median, which is computed many times faster than ImageJ's naive approach.
Without the median, one must set the number of histogram bins to 32 to get
similar overall performance to ImageJ's RankFilter. With the median, even
with 256 bins the integral approach is faster!

The beautiful part is that the integral histogram is computed once,
and then it is possible to iterate, with the same cost each time, for
different radii.

See the TestHistograms.comparePerformance method.